backend/requirements_dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements_dev.in
#
aioredis==1.3.1           # via channels-redis
amqp==2.5.2               # via kombu
appdirs==1.4.3            # via black
asgiref==3.2.7            # via channels, channels-redis, daphne, django
astroid==2.3.3            # via pylint, pylint-celery
async-timeout==3.0.1      # via aioredis
attrs==19.3.0             # via automat, black, flake8-bugbear, pytest, service-identity, twisted
autobahn==20.4.1          # via daphne
automat==20.2.0           # via twisted
bandit==1.6.2             # via -r requirements_dev.in
billiard==3.6.3.0         # via celery
black==19.10b0            # via flake8-strict
celery==4.4.2             # via -r requirements.in
certifi==2020.4.5.1       # via requests
cffi==1.14.0              # via cryptography
channels-redis==2.4.2     # via -r requirements.in
channels==2.4.0           # via -r requirements.in, channels-redis
chardet==3.0.4            # via requests
click==7.1.1              # via black, pip-tools
constantly==15.1.0        # via twisted
coreapi==2.3.3            # via -r requirements.in, django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
coverage==5.0.4           # via pytest-cov
cryptography==2.9         # via autobahn, pyopenssl, service-identity
daphne==2.4.1             # via channels
defusedxml==0.6.0         # via python3-openid
django-allauth==0.41.0    # via -r requirements.in
django-cors-headers==3.2.1  # via -r requirements.in
django-extensions==2.2.9  # via -r requirements.in
django-extra-fields==2.0.3  # via -r requirements.in
django-filter==2.2.0      # via -r requirements.in
django-health-check==3.12.1  # via -r requirements.in
django-redis-cache==2.1.0  # via -r requirements.in
django-rest-auth==0.9.5   # via -r requirements.in
django-rest-swagger==2.2.0  # via -r requirements.in
django==3.0.5             # via -r requirements.in, channels, django-allauth, django-cors-headers, django-extra-fields, django-filter, django-health-check, django-rest-auth, djangorestframework, drf-nested-routers
djangorestframework==3.11.0  # via -r requirements.in, django-extra-fields, django-rest-auth, django-rest-swagger, drf-nested-routers
drf-nested-routers==0.91  # via -r requirements.in
enum-compat==0.0.3        # via flake8-strict
factory-boy==2.12.0       # via -r requirements_dev.in, mimesis-factory
faker==4.0.2              # via factory-boy
flake8-bugbear==20.1.4    # via -r requirements_dev.in
flake8-comprehensions==3.2.2  # via -r requirements_dev.in
flake8-future-import==0.4.6  # via -r requirements_dev.in
flake8-per-file-ignores==0.8.1  # via -r requirements_dev.in
flake8-strict==0.2.1      # via -r requirements_dev.in
flake8==3.6.0             # via -r requirements_dev.in, flake8-bugbear, flake8-comprehensions, flake8-future-import, flake8-per-file-ignores, flake8-strict
gitdb==4.0.2              # via gitpython
gitpython==3.1.0          # via bandit
hiredis==1.0.1            # via aioredis
hyperlink==19.0.0         # via twisted
idna==2.9                 # via hyperlink, requests, twisted
incremental==17.5.0       # via twisted
isort==4.3.21             # via pylint
itypes==1.1.0             # via coreapi
jinja2==2.11.1            # via coreschema
kombu==4.6.8              # via celery
lazy-object-proxy==1.4.3  # via astroid
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8, pylint
mimesis-factory==1.1.0    # via -r requirements_dev.in
mimesis==4.0.0            # via -r requirements_dev.in, mimesis-factory
more-itertools==8.2.0     # via pytest
msgpack==0.6.2            # via channels-redis
oauthlib==3.1.0           # via requests-oauthlib
openapi-codec==1.3.2      # via django-rest-swagger
packaging==20.3           # via pytest
pathmatch==0.2.1          # via flake8-per-file-ignores
pathspec==0.7.0           # via black
pbr==5.4.5                # via stevedore
pillow==7.1.1             # via -r requirements.in
pip-tools==4.5.1          # via -r requirements_dev.in
pluggy==0.13.1            # via pytest
psutil==5.7.0             # via -r requirements.in
psycopg2==2.8.5           # via -r requirements.in
py==1.8.1                 # via pytest
pyasn1-modules==0.2.8     # via service-identity
pyasn1==0.4.8             # via pyasn1-modules, service-identity
pycodestyle==2.4.0        # via flake8
pycparser==2.20           # via cffi
pyenchant==3.0.1          # via -r requirements_dev.in
pyfakefs==4.0.2           # via -r requirements_dev.in
pyflakes==2.0.0           # via flake8
pyhamcrest==2.0.2         # via twisted
pylint-args==0.0.5        # via -r requirements_dev.in
pylint-celery==0.3        # via -r requirements_dev.in
pylint-django==2.0.14     # via -r requirements_dev.in
pylint-plugin-utils==0.6  # via pylint-celery, pylint-django
pylint==2.4.4             # via -r requirements_dev.in, pylint-args, pylint-celery, pylint-django, pylint-plugin-utils
pyopenssl==19.1.0         # via twisted
pyparsing==2.4.7          # via packaging
pytest-asyncio==0.10.0    # via -r requirements_dev.in
pytest-cov==2.8.1         # via -r requirements_dev.in
pytest-django==3.9.0      # via -r requirements_dev.in
pytest==5.4.1             # via pytest-asyncio, pytest-cov, pytest-django
python-dateutil==2.8.1    # via faker
python3-openid==3.1.0     # via django-allauth
pytz==2019.3              # via celery, django
pyyaml==5.3.1             # via bandit
redis==3.4.1              # via -r requirements.in, django-redis-cache
regex==2020.4.4           # via black
requests-oauthlib==1.3.0  # via django-allauth
requests==2.23.0          # via coreapi, django-allauth, requests-oauthlib
service-identity==18.1.0  # via twisted
simplejson==3.17.0        # via django-rest-swagger
six==1.14.0               # via astroid, automat, bandit, cryptography, django-extensions, django-redis-cache, django-rest-auth, packaging, pathmatch, pip-tools, pyopenssl, python-dateutil, stevedore
smmap==3.0.1              # via gitdb
sqlparse==0.3.1           # via django
stevedore==1.32.0         # via bandit
text-unidecode==1.3       # via faker
toml==0.10.0              # via black
twisted[tls]==20.3.0      # via daphne
txaio==20.4.1             # via autobahn
typed-ast==1.4.1          # via black
typing==3.7.4.1           # via pathmatch
uritemplate==3.0.1        # via coreapi
urllib3==1.25.8           # via requests
vine==1.3.0               # via amqp, celery
wcwidth==0.1.9            # via pytest
wrapt==1.11.2             # via astroid
yapf==0.29.0              # via -r requirements_dev.in
zope.interface==5.0.2     # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
